time and sleep

+++ kernel/main.c
@@ -45,13 +45,13 @@ volatile unsigned *irq_ptr;
 
  void sleep (unsigned sleep_time)
 {
-    unsigned wake_up_time = sys_time + sleep_time;
+    unsigned long wake_up_time = sys_time + sleep_time;
 
     while(1)
     {
        if (sys_time > wake_up_time)
             break;
- //     printf("sleeping %lu\n", sys_time);
+      //printf("sleeping %lu until %lu\n", sys_time,wake_up_time);
     }
 }
